feat: recalculate total price when cart gets updated
