+++ src/pages/cart/logic/remove-product-from-cart.ts
@@ -2,6 +2,7 @@ import getActiveCart from '../../../api/get-active-cart';
 import updateCart from '../../../api/update-cart';
 import useToken from '../../../services/use-token';
 import { Cart, CartProduct, UpdateCartData } from '../../../types/cart';
+import updateTotalPrice from './update-total-rice';
 
 export default async function removeProductFromCart(event: Event, cartItem: CartProduct) {
   const token = await useToken.access.get();
@@ -21,6 +22,7 @@ export default async function removeProductFromCart(event: Event, cartItem: Cart
       const cartProductNode = event.target.closest('.cart__product');
       if (updatedCart && cartProductNode instanceof HTMLElement) {
         cartProductNode.remove();
+        updateTotalPrice(updatedCart.totalPrice.centAmount);
       }
     }
   }
