+++ src/pages/cart/logic/update-total-rice.ts
@@ -1,0 +1,6 @@
+export default function updateTotalPrice(totalPriceValue: number) {
+  const totalPrice = document.querySelector('.total');
+  if (totalPrice instanceof HTMLElement) {
+    totalPrice.innerText = `Total: € ${totalPriceValue / 100}`;
+  }
+}
